requirements.txt
requests
kafka-python
pymongo
pymongo[srv]
certifi